Python Backend/paannk/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
motor
google-genai
